refactor(ui): split TypingTestPane into window hook and panel components
